Move cost metric to its own file
